# go/cmd/mcp-server/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST access to the Safecast radiation monitoring dataset — 200M+ measurements from citizen scientists worldwide. All data is CC0-licensed and read-only. Powered by PostgreSQL+PostGIS with a Safecast REST API fallback.",
        "title": "Safecast API",
        "contact": {
            "name": "Safecast",
            "url": "https://safecast.org"
        },
        "license": {
            "name": "CC0 1.0 Universal",
            "url": "https://creativecommons.org/publicdomain/zero/1.0/"
        },
        "version": "1.0"
    },
    "host": "vps-01.safecast.jp",
    "basePath": "/api",
    "paths": {
...
        "/track/{id}": {
            "get": {
                "description": "Retrieves radiation measurements recorded during a specific bGeigie drive. Use GET /api/tracks to find track IDs first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "historical"
                ],
                "summary": "Get all measurements from a specific track",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Track identifier (e.g. 8eh5m1)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Start marker ID for filtering",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "End marker ID for filtering",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 200,
                        "description": "Maximum number of results (1 to 10000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Measurements for the track",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tracks": {
            "get": {
                "description": "Lists bGeigie Import tracks (bulk radiation measurement drives). Each track represents measurements from a single bGeigie session. Can filter by year and optionally month.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "historical"
                ],
                "summary": "Browse bGeigie measurement tracks",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by year (2000–2100)",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by month (1–12, requires year)",
                        "name": "month",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Maximum number of results (1 to 50000)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Track list with count and filter metadata",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
...
}